fix: update the timeout logic to consider also the checkFunction duration
Before this commit, the number of retries was calculated ignoring the checkFunction duration.

As a result, a 5s timeout with a 0.5s interval resulted in 10 retries.
If the checkFunction takes 10 seconds, 10 retries mean waiting for 100 seconds, violating the 5s timeout.

This commit fixes the problem by checking, after each checkFunction invocation, what is the elapsed time and stops retrying in case the timeout is over.

BREAKING CHANGE: The timeout is now respected even if checkFunction takes a long time. As a result,
you could face that your checkFunction runs less times.

fix #464